refactor(sdk-crypto) - RoomKey Sharing | extract share module
